fix: 🐛 Code review. Move dicomLoaderService to core
Moved dicomLoaderService to ohif/core and localFileLoaders to a specific
folder.
