+++ extensions/cornerstone/src/handleSegmentationStorage.js
@@ -5,7 +5,7 @@ import OHIF from '@ohif/core';
 import cornerstone from 'cornerstone-core';
 import cornerstoneTools from 'cornerstone-tools';
 import DICOMWeb from '@ohif/core/src/DICOMWeb';
-import DicomLoaderService from '../../../platform/viewer/src/lib/dicomLoaderService';
+import DicomLoaderService from '@ohif/core/src/utils/dicomLoaderService';
 
 const { StackManager } = OHIF.utils;
