+++ extensions/dicom-pdf/src/OHIFDicomPDFViewport.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import OHIF from '@ohif/core';
 import OHIFComponentPlugin from './OHIFComponentPlugin.js';
 import DicomPDFViewport from './DicomPDFViewport';
-import DicomLoaderService from '../../../platform/viewer/src/lib/dicomLoaderService';
+import DicomLoaderService from '@ohif/core/src/utils/dicomLoaderService';
 
 const { DICOMWeb } = OHIF;
