+++ platform/core/src/utils/dicomLoaderService.js
