+++ platform/viewer/src/lib/localFileLoaders/dicomFileLoader.js
