+++ platform/viewer/src/lib/localFileLoaders/fileLoader.js
