+++ platform/viewer/src/lib/localFileLoaders/fileLoaderService.js
@@ -1,4 +1,5 @@
 import cornerstoneWADOImageLoader from 'cornerstone-wado-image-loader';
+import FileLoader from './fileLoader';
 import PDFFileLoader from './pdfFileLoader';
 import DICOMFileLoader from './dicomFileLoader';
